backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
motor==3.7.0
python-dotenv==1.0.1
openai==1.65.2
pydantic-settings==2.8.1